Switch to buildpackless builder
The integration tests were failing because of buildpack version conflicts in the builder. We can use the buildpackless builder which has no buildpacks in it. This should not change the behavior as we are currently pulling in a complete list of buildpacks which are required to run the integration test.

Signed-off-by: Daniel Mikusa <dmikusa@vmware.com>